split survey questions into a separate file
